Make `BasePolaritsMetaStoreManagerTest` and `(Base)ResolverTest` reusable
Moves the test cases into the `Base*` classes and make sure the classes can be reused by other persistence implementations.
